Refactor ContacUs: extract ContactInformation component
Move ContacUs.jsx into its own folder and extract the map/contact info
block into a dedicated ContactInformation component.
